move redshift helper methods to mixin
* so they're available from mosviz & cubeviz, as well as specviz
